Renames useHasMounted hook filename
